fix(App): Moving CSS to its own directory
