+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import './Header.css';
+import './css/Header.css';
 import {Container, Image, Menu} from 'semantic-ui-react';
 
 const Header = () => {
